+++ spec/fuzzing/extensions/string_spec.rb
@@ -21,52 +21,10 @@ describe String do
   describe "generate" do
     subject { described_class }
 
-    it "should generate Strings from CharSets" do
-      strings = subject.generate(:lowercase_hexadecimal, :numeric).to_a
-
-      strings.grep(/^[0-9a-f][0-9]$/).should == strings
-    end
-
-    it "should generate Strings from lengths of CharSets" do
+    it "should generate Strings from a template" do
       strings = subject.generate([:numeric, 2]).to_a
 
       strings.grep(/^[0-9]{2}$/).should == strings
-    end
-
-    it "should generate Strings from varying lengths of CharSets" do
-      strings = subject.generate([:numeric, 1..2]).to_a
-
-      strings.grep(/^[0-9]{1,2}$/).should == strings
-    end
-
-    it "should generate Strings from custom CharSets" do
-      strings = subject.generate([%w[a b c], 2]).to_a
-
-      strings.grep(/^[abc]{2}$/).should == strings
-    end
-
-    it "should generate Strings containing known Strings" do
-      strings = subject.generate('foo', [%w[a b c], 2]).to_a
-
-      strings.grep(/^foo[abc]{2}$/).should == strings
-    end
-
-    it "should raise a TypeError for non String, Symbol, Enumerable CharSets" do
-      lambda {
-        subject.generate([Object.new, 2]).to_a
-      }.should raise_error(TypeError)
-    end
-
-    it "should raise an ArgumentError for unknown CharSets" do
-      lambda {
-        subject.generate([:foo_bar, 2]).to_a
-      }.should raise_error(ArgumentError)
-    end
-
-    it "should raise a TypeError for non Integer,Array,Range lengths" do
-      lambda {
-        subject.generate([:numeric, 'foo']).to_a
-      }.should raise_error(TypeError)
     end
   end